Refine details in Test cases
