+++ src/test/java/com/zoomcare/candidatechallenge/CandidateChallengeApplicationTest.java
@@ -1,4 +1,4 @@
-package test;
+package com.zoomcare.candidatechallenge;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -7,7 +7,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class CandidateChallengeTest {
+public class CandidateChallengeApplicationTest {
 
 	@Test
 	public void contextLoads() {
